backend/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
websockets==15.0.1

# Database
sqlalchemy==2.0.40
alembic==1.15.2
aiosqlite==0.21.0

# Auth
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# AI
anthropic==0.52.0

# Validation
pydantic==2.11.1
pydantic-settings==2.8.1
email-validator==2.2.0

# Utils
python-dotenv==1.1.0
python-dateutil==2.9.0

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
httpx==0.28.1
factory-boy==3.3.3
